+++ backend/tests/Functional/Controller/User/AuthControllerTest.php
@@ -436,7 +436,7 @@ class AuthControllerTest extends WebTestCase
                 HttpClientFactoryInterface::class => new HttpClientFactory($this->client),
                 LoggerInterface::class => $this->log,
             ],
-            [['NEUCORE_EVE_SCOPES', 'read-this'], ['NEUCORE_EVE_DATASOURCE', 'tranquility']],
+            [['NEUCORE_EVE_SCOPES', 'read-this']],
         );
         $this->assertSame(302, $response->getStatusCode());
 
@@ -478,7 +478,7 @@ class AuthControllerTest extends WebTestCase
             null,
             null,
             [HttpClientFactoryInterface::class => new HttpClientFactory($this->client)],
-            [['NEUCORE_EVE_SCOPES', 'read-this   and-this'], ['NEUCORE_EVE_DATASOURCE', 'tranquility']],
+            [['NEUCORE_EVE_SCOPES', 'read-this   and-this']],
         );
         $this->assertSame(302, $response->getStatusCode());
 
@@ -514,7 +514,7 @@ class AuthControllerTest extends WebTestCase
                 HttpClientFactoryInterface::class => new HttpClientFactory($this->client),
                 LoggerInterface::class => $this->log,
             ],
-            [['NEUCORE_EVE_SCOPES', 'read-this'], ['NEUCORE_EVE_DATASOURCE', 'tranquility']],
+            [['NEUCORE_EVE_SCOPES', 'read-this']],
         );
         $this->assertSame(302, $response->getStatusCode());
 
@@ -554,7 +554,7 @@ class AuthControllerTest extends WebTestCase
             null,
             null,
             [HttpClientFactoryInterface::class => new HttpClientFactory($this->client)],
-            [['NEUCORE_EVE_SCOPES', 'read-this'], ['NEUCORE_EVE_DATASOURCE', 'tranquility']],
+            [['NEUCORE_EVE_SCOPES', 'read-this']],
         );
         $this->assertSame(302, $response->getStatusCode());
 
@@ -602,7 +602,7 @@ class AuthControllerTest extends WebTestCase
                 HttpClientFactoryInterface::class => new HttpClientFactory($this->client),
                 LoggerInterface::class => $this->log,
             ],
-            [['NEUCORE_EVE_SCOPES', 'read-this'], ['NEUCORE_EVE_DATASOURCE', 'tranquility']],
+            [['NEUCORE_EVE_SCOPES', 'read-this']],
         );
         $this->assertSame(302, $response->getStatusCode());
 
@@ -646,7 +646,7 @@ class AuthControllerTest extends WebTestCase
                 HttpClientFactoryInterface::class => new HttpClientFactory($this->client),
                 LoggerInterface::class => $this->log,
             ],
-            [['NEUCORE_EVE_SCOPES', 'read-this'], ['NEUCORE_EVE_DATASOURCE', 'tranquility']],
+            [['NEUCORE_EVE_SCOPES', 'read-this']],
         );
         $this->assertSame(302, $response->getStatusCode());
 
@@ -686,7 +686,6 @@ class AuthControllerTest extends WebTestCase
             null,
             null,
             [HttpClientFactoryInterface::class => new HttpClientFactory($this->client)],
-            [['NEUCORE_EVE_DATASOURCE', 'tranquility']],
         );
         $this->assertSame(302, $response->getStatusCode());
 
@@ -725,7 +724,6 @@ class AuthControllerTest extends WebTestCase
             null,
             null,
             [HttpClientFactoryInterface::class => new HttpClientFactory($this->client)],
-            [['NEUCORE_EVE_DATASOURCE', 'tranquility']],
         );
         $this->assertSame(302, $response->getStatusCode());