+++ backend/tests/Unit/Service/EsiDataTest.php
@@ -64,7 +64,6 @@ class EsiDataTest extends TestCase
         $this->log->pushHandler(new TestHandler());
 
         $this->config = new Config(['eve' => [
-            'datasource' => '',
             'esi_host' => '',
             'esi_compatibility_date' => '',
         ]]);